django_api_to_upload_file_into_local_database_and_s3Bucket/image_upload_api/imga_upload/app/serializer.py
from rest_framework import serializers
from .models import file_model

class file_serializer(serializers.ModelSerializer):

    class Meta():
        model = file_model
        fields = ('file','key','timestamp')