django_api_to_upload_file_into_local_database_and_s3Bucket/image_upload_api/imga_upload/requirement.txt
Django==1.11.17
django-rest-swagger==2.2.0
djangorestframework==3.9.0
virtualenv==16.1.0
mysqlclient
pillow
